+++ src/main/java/framework/com/giisoo/framework/common/Message.java
@@ -36,7 +36,7 @@ public class Message extends Bean {
 	public static final String FLAG_DONE = "done";
 
 	String id;
-	int uid;
+	long uid;
 	int from_uid;
 
 	String subject;
@@ -50,7 +50,7 @@ public class Message extends Bean {
 		return id;
 	}
 
-	public int getUid() {
+	public long getUid() {
 		return uid;
 	}
 
@@ -87,7 +87,7 @@ public class Message extends Bean {
 		return flag;
 	}
 
-	private static String getID(int uid, String clazz, String subject) {
+	private static String getID(long uid, String clazz, String subject) {
 		return UID.id(uid, clazz, subject, System.currentTimeMillis());
 	}
 
@@ -106,7 +106,7 @@ public class Message extends Bean {
 	 *            the from
 	 * @return the int
 	 */
-	public static int create(int uid, Class<?> clazz, String subject,
+	public static int create(long uid, Class<?> clazz, String subject,
 			String refer, int from) {
 
 		String id = getID(uid, null, subject);
@@ -136,7 +136,7 @@ public class Message extends Bean {
 	 *            the from
 	 * @return the int
 	 */
-	public static int create(int uid, Class<?> clazz, String subject,
+	public static int create(long uid, Class<?> clazz, String subject,
 			String where, Object[] args, int from) {
 
 		JSONObject jo = new JSONObject();
@@ -160,7 +160,7 @@ public class Message extends Bean {
 	 *            the from
 	 * @return the int
 	 */
-	public static int create(int uid, String subject, String body, int from) {
+	public static int create(long uid, String subject, String body, int from) {
 		String id = getID(uid, null, subject);
 		return Bean.insert(
 				V.create("uid", uid).set("subject", subject).set("body", body)
@@ -181,7 +181,7 @@ public class Message extends Bean {
 	 *            the v
 	 * @return the int
 	 */
-	public static int create(int uid, String id, V v) {
+	public static int create(long uid, String id, V v) {
 		if (!Bean.exists("uid=? and id=?", new Object[] { uid, id },
 				Message.class)) {
 			return Bean.insert(v.set("uid", uid).set("id", id), Message.class);
@@ -201,7 +201,7 @@ public class Message extends Bean {
 	 *            the n
 	 * @return the beans
 	 */
-	public static Beans<Message> load(int uid, int s, int n) {
+	public static Beans<Message> load(long uid, int s, int n) {
 		return Bean.load("uid=?", new Object[] { uid },
 				"order by created desc", s, n, Message.class);
 	}
@@ -219,7 +219,7 @@ public class Message extends Bean {
 	 *            the n
 	 * @return the beans
 	 */
-	public static Beans<Message> load(int uid, W w, int s, int n) {
+	public static Beans<Message> load(long uid, W w, int s, int n) {
 		w.and("uid", uid);
 
 		return Bean.load(w.where(), w.args(),
@@ -240,7 +240,7 @@ public class Message extends Bean {
 	 *            the n
 	 * @return the beans
 	 */
-	public static Beans<Message> load(int uid, String clazz, int s, int n) {
+	public static Beans<Message> load(long uid, String clazz, int s, int n) {
 		return Bean.load("uid=? and clazz=?", new Object[] { uid, clazz },
 				"order by created desc", s, n, Message.class);
 	}
@@ -254,7 +254,7 @@ public class Message extends Bean {
 	 *            the w
 	 * @return the list
 	 */
-	public static List<Group> group(int uid, W w) {
+	public static List<Group> group(long uid, W w) {
 		w.and("uid", uid, W.OP_EQ);
 		return Bean.load("tblmessage", new String[] { "clazz", "count(*) t" },
 				w.where(), w.args(), "group by clazz", -1, -1, Group.class,
@@ -270,7 +270,7 @@ public class Message extends Bean {
 	 *            the flag
 	 * @return the list
 	 */
-	public static List<Group> group(int uid, String flag) {
+	public static List<Group> group(long uid, String flag) {
 		return Bean.load("tblmessage", new String[] { "clazz", "count(*) t" },
 				"uid=? and flag=?", new Object[] { uid, flag },
 				"group by clazz", -1, -1, Group.class, null);
@@ -287,7 +287,7 @@ public class Message extends Bean {
 	 *            the v
 	 * @return the int
 	 */
-	public static int update(int uid, String id, V v) {
+	public static int update(long uid, String id, V v) {
 		return Bean.update("uid=? and id=?", new Object[] { uid, id }, v,
 				Message.class);
 	}
@@ -312,7 +312,7 @@ public class Message extends Bean {
 	 *            the id
 	 * @return the int
 	 */
-	public static int delete(int uid, String id) {
+	public static int delete(long uid, String id) {
 		return Bean.delete("uid=? and id=?", new Object[] { uid, id },
 				Message.class);
 	}
@@ -324,7 +324,7 @@ public class Message extends Bean {
 	protected void load(ResultSet r) throws SQLException {
 
 		id = r.getString("id");
-		uid = r.getInt("uid");
+		uid = r.getLong("uid");
 		from_uid = r.getInt("from_uid");
 
 		subject = r.getString("subject");
@@ -374,7 +374,7 @@ public class Message extends Bean {
 	 *            the id
 	 * @return the message
 	 */
-	public static Message load(int uid, String id) {
+	public static Message load(long uid, String id) {
 		return Bean.load("uid=? and id=?", new Object[] { uid, id },
 				Message.class);
 	}
